moved function to utils (re)move helper functions to another module
